New pandda

+++ pandda_gemmi/scratch/event_model/cluster.py
@@ -72,7 +72,7 @@ class ClusterDensityDBSCAN:
                 continue
             events[j] = Event(
                 high_z_pos_array[clusters == cluster_num, :],
-                high_z_point_array[clusters == cluster_num, :],
+                high_z_point_array[clusters == cluster_num, :].astype(np.int),
                 0.0,
             )
             j +=1
